Standardize Connector ID Type to int64
LiveReview Pre-Commit Check: ran (iter:1, coverage:0%)

+++ internal/provider_input/azuredevops/azuredevops_auth.go
@@ -93,7 +93,7 @@ func FindIntegrationTokenForAzureDevOpsOrg(db *sql.DB, orgURL string) (*Integrat
 }
 
 // FindIntegrationTokenByConnectorID finds the Azure DevOps integration token by connector ID.
-func FindIntegrationTokenByConnectorID(db *sql.DB, connectorID int) (*IntegrationToken, string, error) {
+func FindIntegrationTokenByConnectorID(db *sql.DB, connectorID int64) (*IntegrationToken, string, error) {
 	if db == nil {
 		return nil, "", fmt.Errorf("database connection is nil")
 	}
